pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.5",
    "paho-mqtt>=2.1.0",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "plotly>=6.5.0",
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
]
